Abstract payment method components to common and use throughout app
